fix header

+++ diffdrive_arduino/include/diffdrive_arduino/arduino_comms.h
@@ -20,7 +20,7 @@ public:
   void setup(const std::string &serial_device, int32_t baud_rate, int32_t timeout_ms);
   void sendEmptyMsg();
   void readEncoderValues(int &val_1, int &val_2);
-  void setMotorValues(int val_1, int val_2);
+  void setMotorValues(double val_1, double val_2);
   void setCmdValues(double val_l, double val_r);
   
   void setPidValues(float k_p, float k_d, float k_i, float k_o);
